project2/P2 Deliverable 1/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
markdown-it-py==4.0.0
mdurl==0.1.2
openai==2.5.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-dotenv==1.1.1
regex==2025.9.18
requests==2.32.5
rich==14.2.0
sniffio==1.3.1
tiktoken==0.12.0
tinytroupe==0.0.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
